Virtual machine fix to allow method invocations in argument expressions
There was a virtual machine bug that assumed some level of fixedness in
the evaluation of parameters. This made it impossible for “open-ended”
expressions, like method invocations, to appear as method parameters.
Most of the fixes were in RTExpression but it also required a couple of
hooks in the virtual machine.

System.out.print().print().println() therefore now works.

NOTE for the future: We need to get some basic machine-stepping-cycle
processing centralised in the run-time environment. This fix, which
required a loop around the stepping through the execution sequence of
parameter evaluation, takes too much execution out of the base machine
cycle. Instead of running-it-from-the-top we need to tap into the cycle
at the program-counter level. This will be required to supporting
tracing and breakpoints in the future.

Also started coding up some stuff in TestRunner.java (and a bit in
TextEditorGUI.java) to be able to check against gold results for
regression tests.

Last, fixed parsing trace output for FOR statements to bring it up to
snuff with the most recent FOR generalisations.