Removed nextCode processing from statement nodes / fixed re-run bug
The “run” method on object virtual machine instructions itself has
always returned the address of the next instruction but the virtual
machine has persisted in using an additional “nextCode” inquiry to
re-fetch the address. The code has been cleaned up to use only the
returned next address. Main changes are in RTExpression.java.

Some test cases were found not to be re-runnable; the problem was
traced to the way constant bindings work. Fixed with a dup call on
RTConstants in RTExpression.java.

Minor cosmetic changes elsewhere.