More generics parsing foundations
These changes update the approach to parsing templates, doing basic
semantic analysis of templates, and generating template instantiations.
It’s a bit ad-hoc and is designed not to interfere with generic-free
code.

This is a checkpoint that seems to have a lot more of the necessary
architecture in place and which seems sound.

Where I left things, all tests pass, but much of the template method
code isn’t making it through to the RT* world.