First instalment on Role vectors
Trygve’s implementation binds a vector to a Role to give all vector
elements role behaviour. However, each one should play a distinct
“instance” of the Role. And we should be able to assign individual role
players of a vector to play a vector Role, as well as treat the vector
object as a Role-player itself. This begs a syntax to distinguish
vector Role-players. This branch is a first cut at this change. All
standard tests pass. Individual Role assignment has been somewhat
tested. Assignment of a vector to a Role vector name has not been
tested. Assignment of a vector to a scalar role that supports vector
methods has not been tested and may be made illegal.