Introducing access qualifier enforcement
The rudimentary logic for access qualifier enforcement has been added.
Only private and public are supported.

There were also some edits to make error message formats more
consistent, to add a test for access control, to add a version number
to the Copyright notice, to make the test leave the interface in a bit
more elegant shape in the event of test failure, and fixed some
declaration bugs.