Check for method of same signature in both Role and its Roleplayer's class
The environment now issues a warning if the programmer assigns an
object to a Role where the object’s class contains a method whose
signature is the same as that of a Role signature. It’s not airtight
(the object may actually be a derived class object, typed to a base
class identifier…) but it may ward off intentional but ignorant
behaviour.