Fixed right-associativity abelian bug
The new grammar introduced a bug in the way expressions are evaluated.
Normally, a * b * c * d should be evaluated left-to-right but it was
being evaluated backwards. The same is true for SUMOPS as for MULOPS. I
used a stack to reverse the order of evaluation. It all works now.

All tests pass.