Automatic promotion logic
Added logic to promote integers to doubles (with a warning). Because
this requires more compile-time knowledge closer to run time, it was a
big change.

Also made it easier to manage the menu when loading tests from the web.

There was some amount of cleanup and adding of “final” modifiers.