Changes to test runner to explore Windows compatibility problem
The test runs on Windows don’t match the gold outputs. The problem
seems to lie in special characters (\012, \014, \015, etc.) that are
generated by the Mac and not Windows, or vice versa. Mapping all of
them to null for the test comparison. Also added some octal dumps to
see what is going on.