Andreas bug fixes; superclass constructor orchestration
Fixed a couple of bugs from Andreas. List<T>::get was completely
broken: the signature was wrong in ListClass.java and template return
types were not properly being handled. Also a problem related to
QualifiedIdentifier, because Pass1Listener has not yet established the
type of “this” so lhs_ checks in several productions stumbled on the
type of “this.” Just put in error stumbling logic and do it right in
passes 2-4.

Added logic that makes it possible to invoke the (single) base class
constructor from a derived class constructor. It must be the first
statement in the derived class constructor. It overrides the default
behaviour of implicitly calling the default Base class constructor if
it exists.

All tests pass… I can swear that simpletemplate.k failed last night but
cannot reproduce the failure.

New bug discovered; caputred in andreas3.k.