Simple assert; doc upgrades; bug fixes
I fixed several bugs Andreas reported: the use of a naked NEW operator
(which he was using for his Assert class); the inability to use an
x.toString() inside a println() argument; and the inability to declare
an empty role definition list on a stage prop.

Also made some minor additions to the doc and to the README with
Andreas’s nice simple bring-up instructions.

I put a VERY rudimentary assert facility into class Object. This led to
some changes to the “halt propagation” architecture in the run time.

Nice example from Andreas was added to the tests:
tests/luhnvalidation.k.