Several more Andreas bug fixes
Fixed a bug where shadowing detection was asserting instead of printing
a warning.

Fixed a problem where it couldn’t find a declaration because it wasn’t
propagating up scopes. This included a cleanup in nearestEnclosing
MethodScope around.

Fixed some scope management logic that was associating a method
declaration with a block. Searching for the nearest enclosing method
would yield the block… Blocks and for loops now have a null associated
declaration.

Added logical or, logical and, and logical XOR processing to the
run-time.

Additional cleanup and corrections to last fix for stage props and
roles.

Fixed a bug where the machine initialization was creating two distinct
boolean types…

All tests run.