Fixed bug with calling Role method from within a block
The lookup logic wasn’t working if a Role method was called from with a
block or loop of another Role method. Changed the lookup logic to
percolate up the chain looking for “this” and “current$context”.

Also fixed a residual state bug, where the run-time environment was
holding on to the “lastExpression” evaluation from the previous run…
This would in general cause the ensuing run to crash when the garbage
collector tried to make things at peace for the old object, all of
whose friends had disappeared…

All tests pass.