Conjectural bug fix for current$context overwrite
There was a bug where having a Role argument to a Context constructor
was causing the compiler to look up the Role in the wrong Context. It
turns out that current$context was being overwritten with the attempt
of setting up a temporary variable in the current activation record, to
be pushed on the stack, and then to be pulled into the activation
record of the called routine. It turns out that the temporary entry in
the calling activation record seems unnecessary…

All tests pass but keep an eye on this one…