"Naked new" bug
Matt found a bug that “naked new” invocations created zombie objects.
Added a facility to detect whether these objects were consumed and, if
not, to bind them to an identifier in the current scope so scope
reclamation cleaned them up.

This took a bit of work at the global scope level and the boundary
condition of program startup… Added code to populate to top activation
record with object declarations at global scope…