Assertions from Roles
I added lots of fake-outs in a lot of the compile- and run-time lookup
code, so that when searching for a method in a Role scope, it goes on
to the class Object scope if it doesn’t find it. Seems fair. Every
Role-player will have Object as a base class. So now assertions inside
of Role methods get found.

All tests pass.