1.5.13: Better task / window integrity support; println of <null> issues a line feed
Per Rune’s request in #86, a println of null now occurs on its own line.

Lots of little updates to deal with program interrupt, button display
status, and making sure the proper stuff gets cleaned up.

Better error handling in the spirit of #82, for a lot of the Panel and
Frame APIs.

All tests pass.