1.5.17:  Minor fix for "[trygve] Map<K,V> containsKey not working with String keys (#88)" and new type system
There is a small bug fix in this push, but the big news is that the
type system has been extensively polished. It has been both tightened
up (being attentive to alignment between, e.g., Roles and Interface or
between List<X> and List<Y>) but also made more liberal (much more
duck-like).

This is a major change and it may precipitate some bugs.

But all tests currently pass.