1.6.16: Fixed Role / Role vector anomaly
It had become legal in some cases to invoke a Role vector method using
only the Role name, with no subscript. Undeterminate run-time behavior.
Fixed so that this usually is not allowed — however, within a Role, it
uses the current RolePlayer and the syntax is legal.

Added logic to null out arrays.

Minor document corrections.

Added clarifying arrows on “Open” and “Save” buttons on GUI.