Initial insertion of debugger foundation code
This is just the initial commit of basic debugger support code. It is
known not yet to work. Explorers are welcome to experiment with it. It
currently will not work if you edit the text buffer during the edit /
debug cycle with the debugger open. Rest of the design coming soon.