3.11: Interim push of debugger enhancement
Fixed a few debugger bugs: printing integers, printing the enclosing
scopes in a method, and fixing a bug in the speed optimization that was
causing statements to be left out. Also redid the window to use a
GroupLayout instead of the wimpy FlowLayout. A clear button for the
debugger transcript lot. I also added some interim stuff for filtering
but it does not yet work.