3.13: Debugger cursor fix, etc.
The main fix is to supply a cursor to the debugger windows when it
mysteriously disappears, maybe because multiple threads are writing in
the same window. Also added the helloworld.k example, and added a very
basic CONTRIBUTING.md file. Last, several updates to trygve1.docx,
including a basic description of the debugger.