trygve 3.14: Fix issue #65
trygve was complaining if a Context instance script had a formal
parameter with a Role type. This should be discouraged. Changed things
so that a warning is issued upon finding this — unless the script is a
constructor. It’s good documentation to type the Role-player arguments
when creating a Context, and because it’s a constructor there’s no
danger that the script is called from elsewhere within the same Context
object.

Added a new test, issue65.k, that elicits the warning.