Per issue 133.
The test file file:examples/keypad.k elicits a fatal assertion (assert false) if invoked, when the cursor is exercised vigorously over the example's keypad. The assertion occurs in RTObjectCommon>>decrementReference count (see tag: REF_COUNT_TAG_BUG_1 in the code).

The problem APPEARS to be a mismanagement in the reference count of qualified identifiers. A negative reference count elicited the assertion.

In RTExpression.java, RTQualifiedIdentifierPart2>>run, we find this fragment:

// At the beginning of the evaluation of this qualified identifier,
// the qualifier was pushed on the stack. Stack pushes increment
// the reference count. Pops do not. At the beginning of this method
// we popped it. Undo the reference count manipulation.

qualifier.decrementReferenceCount();
This is apparently what causes the reference count to go negative.

What exactly is going on is not clear, but it seems like this reference count decrement is inappropriate. If we remove it, testing can no longer elicit the assertion in RTObjectCommon>>decrementReferenceCount.

Both locations have been tagged with REF_COUNT_TAG_BUG_1 in code commentary.