When running the built-in tests, some tests fail quite regularly with what appears to be a gratuitous newline. The extra character often appears after the long underscore separator in test output. Occasionally I have seen a missing newline between System.err output and immediately ensuing System.out output.
The solution appears to be related to the presence of \n\r sequences, either in the Golds or in the test output, as a result of a println() call in the program. De-Microsofting both the Gold and the test output causes the test results to compare favorably to the Golds.